*** empty log message ***

+++ src/sys/os/os.c
@@ -178,7 +178,7 @@ static void *load_image(char *filename)
 	return NULL;
       }
 
-      if (read(f, RVA(imgbase, imghdr->sections[i].virtual_address), imghdr->sections[i].size_of_raw_data) != imghdr->sections[i].size_of_raw_data)
+      if (read(f, RVA(imgbase, imghdr->sections[i].virtual_address), imghdr->sections[i].size_of_raw_data) != (int) imghdr->sections[i].size_of_raw_data)
       {
 	munmap(imgbase, imghdr->optional.size_of_image, MEM_RELEASE);
 	close(f);
